+++ sdk/identity/azure-identity/azure/identity/_credentials/managed_identity.py
@@ -11,6 +11,7 @@ from azure.core.pipeline.policies import (
     ContentDecodePolicy,
     DistributedTracingPolicy,
     HeadersPolicy,
+    HttpLoggingPolicy,
     NetworkTraceLoggingPolicy,
     RetryPolicy,
 )
@@ -25,25 +26,28 @@ except ImportError:
 
 if TYPE_CHECKING:
     # pylint:disable=unused-import
-    from typing import Any, Mapping, Optional, Type
+    from typing import Any, Optional, Type
 
 
 class ManagedIdentityCredential(object):
-    """Authenticates with a managed identity in an App Service, Azure VM or Cloud Shell environment.
+    """Authenticates with an Azure managed identity in any hosting environment which supports managed identities.
 
-    :param str client_id:
-        (optional) client ID of a user-assigned identity. Leave unspecified to use a system-assigned identity.
+    See the Azure Active Directory documentation for more information about managed identities:
+    https://docs.microsoft.com/en-us/azure/active-directory/managed-identities-azure-resources/overview
+
+    Keyword arguments
+        - **client_id** (str): ID of a user-assigned identity. Leave unspecified to use a system-assigned identity.
     """
 
-    def __new__(cls, *args, **kwargs):
+    def __new__(cls, **kwargs):
         if os.environ.get(EnvironmentVariables.MSI_ENDPOINT):
-            return MsiCredential(*args, **kwargs)
-        return ImdsCredential(*args, **kwargs)
+            return MsiCredential(**kwargs)
+        return ImdsCredential(**kwargs)
 
     # the below methods are never called, because ManagedIdentityCredential can't be instantiated;
     # they exist so tooling gets accurate signatures for Imds- and MsiCredential
-    def __init__(self, client_id=None, **kwargs):
-        # type: (Optional[str], Any) -> None
+    def __init__(self, **kwargs):
+        # type: (**Any) -> None
         pass
 
     def get_token(self, *scopes, **kwargs):  # pylint:disable=unused-argument,no-self-use
@@ -69,17 +73,16 @@ class _ManagedIdentityBase(object):
             config.headers_policy,
             config.retry_policy,
             config.logging_policy,
-            DistributedTracingPolicy(),
+            DistributedTracingPolicy(**kwargs),
+            HttpLoggingPolicy(**kwargs),
         ]
         self._client = client_cls(endpoint=endpoint, config=config, policies=policies, **kwargs)
 
     @staticmethod
     def _create_config(**kwargs):
-        # type: (Mapping[str, Any]) -> Configuration
-        """Build a default configuration for the credential's HTTP pipeline.
+        # type: (**Any) -> Configuration
+        """Build a default configuration for the credential's HTTP pipeline."""
 
-        :rtype: :class:`azure.core.configuration`
-        """
         timeout = kwargs.pop("connection_timeout", 2)
         config = Configuration(connection_timeout=timeout, **kwargs)
 
@@ -109,8 +112,8 @@ class _ManagedIdentityBase(object):
 class ImdsCredential(_ManagedIdentityBase):
     """Authenticates with a managed identity via the IMDS endpoint.
 
-    :param config: optional configuration for the underlying HTTP pipeline
-    :type config: :class:`azure.core.configuration`
+    Keyword arguments:
+        - **client_id** (str): ID of a user-assigned identity. Leave unspecified to use a system-assigned identity.
     """
 
     def __init__(self, **kwargs):
@@ -161,18 +164,15 @@ class ImdsCredential(_ManagedIdentityBase):
 class MsiCredential(_ManagedIdentityBase):
     """Authenticates via the MSI endpoint in an App Service or Cloud Shell environment.
 
-    :param config: optional configuration for the underlying HTTP pipeline
-    :type config: :class:`azure.core.configuration`
+    Keyword arguments:
+      - **client_id** (str): ID of a user-assigned identity. Leave unspecified to use a system-assigned identity.
     """
 
-    def __init__(self, config=None, **kwargs):
-        # type: (Optional[Configuration], Mapping[str, Any]) -> None
-        endpoint = os.environ.get(EnvironmentVariables.MSI_ENDPOINT)
-        self._endpoint_available = endpoint is not None
-        if self._endpoint_available:
-            super(MsiCredential, self).__init__(  # type: ignore
-                endpoint=endpoint, client_cls=AuthnClient, config=config, **kwargs
-            )
+    def __init__(self, **kwargs):
+        # type: (**Any) -> None
+        self._endpoint = os.environ.get(EnvironmentVariables.MSI_ENDPOINT)
+        if self._endpoint:
+            super(MsiCredential, self).__init__(endpoint=self._endpoint, client_cls=AuthnClient, **kwargs)
 
     def get_token(self, *scopes, **kwargs):  # pylint:disable=unused-argument
         # type: (*str, **Any) -> AccessToken
@@ -183,7 +183,7 @@ class MsiCredential(_ManagedIdentityBase):
         :raises: :class:`azure.core.exceptions.ClientAuthenticationError`
         """
 
-        if not self._endpoint_available:
+        if not self._endpoint:
             raise ClientAuthenticationError(message="MSI endpoint unavailable")
 
         if len(scopes) != 1: