+++ sdk/storage/azure-storage-blob/tests/test_blob_samples_authentication.py
@@ -89,9 +89,9 @@ class TestAuthSamples(StorageTestCase):
         # Get a token credential for authentication
         from azure.identity import ClientSecretCredential
         token_credential = ClientSecretCredential(
+            self.active_directory_tenant_id,
             self.active_directory_application_id,
-            self.active_directory_application_secret,
-            self.active_directory_tenant_id
+            self.active_directory_application_secret
         )
 
         # Instantiate a BlobServiceClient using a token credential